fix linux compilation

+++ process/process_optimize_mesh.cpp
@@ -3,7 +3,7 @@
 void process_optimize_mesh(process_t& process) {
 	float threshold = 0.001f;
 	int		welded = 0;
-	for (int i = 1; i < process.vertices.size(); ++i) {
+	for (int i = 1; i < (int) process.vertices.size(); ++i) {
 		const auto& v = process.vertices[i];
 		for (int j = 0; j < i; ++j) {
 			const auto vc = v3::sub(process.vertices[j], v);
@@ -21,7 +21,7 @@ void process_optimize_mesh(process_t& process) {
 				}
 				welded++;
 				break;
-			}			
+			}
 		}
 	}
 	std::cout << "x welded " << welded << " point(s)" << std::endl;
